requirements.txt
fastapi
motor
pydantic
certifi
uvicorn
gunicorn
websockets
wrapt
python-dotenv
pyjwt